created TODO.yml
This will automatically create issues from TODO statements in our code. 

This will not automatically add it to the repo project however, additional work using project secrets is required for that. See [TODO to Issue Wiki}(https://github.com/marketplace/actions/todo-to-issue) to possibly add this in the future

Example Use:
```
    def hello_world():
        # TODO Come up with a more imaginative greeting
        #  Everyone uses hello world and it's boring.
        #  labels: enhancement, help wanted
        #  assignees: austinblackmon, theautomelon
        #  milestone: 1
        print('Hello world!')
```